Replace listpeers.channels with listpeerchannels
Renamed the type Channels to ChannelsDomain since it has extra fields added.  ChannelsDto is the response from the network.  This was done so it is clear what comes from the network and what is calculated, and to map fields from listpeerchannels into the previous Channel domain model.  Removed some fields that are not used and not available in listpeerchannels: direction, max_htlc_value_in_flight_msat, htlc_minimum_msat, last_tx_fee_msat.  Other fields were duplicates such as spendable_msatoshi and deleted since spendable_msat exists in listpeerchannels as well as in the original Channels object.

use-http sendRequestToStore is now able to handle multiple rpc methods and it bundles the responses together.  This allows for sending 2+ rpc calls to 1 state, such as for the new channels AppContextType.

Removed the controllers/lightning.ts behavior of checking if listpeers was called and adding node.alias to the response.  This was so that the domain layer would handle setting domain data.

channels was added as an AppContextType.  It combines listpeerchannels and listnodes so that the node aliases can be set to the channels.

Added comment to isCompatibleVersion to explain what it does.

Added Channels component tests.  Update mock store to use channels instead of listChannels.