Fetch and Process Balance Sheet data
This commit uses the `sql` rpc command to fetch the data for the Balance Sheet.

getBalanceSheet() was added to use-http.ts as a standalone function instead of getting it via AppContext so that we can pass parameters more easily to it.  TimeGranularity represents the way we should group the data e.g. Monthly, Weekly, Daily.

The sql needs to be transformed for the balance sheet, this was done in a new file bookkeeper-transform.ts since it takes some effort.  Future transform functions such as for Sats Flow would go in this file.  transformToBalanceSheet() is complex because we are limited in what methods we can use in the `sql` rpc call, so I need to do things like grouping and aggregation of credits - debits.