Display Invoice Rune
This commit adds an invoice rune to the ConnectWallet component.

On application start, we try to read an INVOICE_RUNE from .commando-env.  One will be written to .commando-env when the user clicks to create one, if none exist already.  If one does exist, it gets passed through and shown on the screen.