Display Invoice Rune (#56)
* Display Invoice Rune

This commit adds an invoice rune to the ConnectWallet component.

On application start, we try to read an INVOICE_RUNE from .commando-env.  One will be written to .commando-env when the user clicks to create one, if none exist already.  If one does exist, it gets passed through and shown on the screen.

* Cosmetic updates for Invoice Rune

- Added tooltip for `create new` button
- Added placeholder if the invoice rune in empty
- Changed spinner type to application's UI standard
- Added error detail in toast if creation fails
- Changed to "a bit" slicker svg for Add

---------

Co-authored-by: ShahanaFarooqui <shahana.farooqui@gmail.com>